The great refactoring
