+++ src/misc/string_interning.rs
